requirements.txt
# Procesamiento de datos
pandas==2.0.3
numpy==1.24.3
PyPDF2==3.0.1
pdfplumber==0.10.2
langdetect==1.0.9

# Embedding y vectorización
sentence-transformers==2.2.2
faiss-gpu==1.7.4
scikit-learn==1.2.2
umap-learn==0.5.3

# LLM y generación
torch==2.0.1
transformers==4.35.0
accelerate==0.23.0
bitsandbytes==0.41.1
optimum==1.12.0

# UI y aplicación
streamlit==1.28.0
streamlit-chat==0.1.1
plotly==5.14.1
matplotlib==3.7.1
seaborn==0.12.2

# Logging y monitoreo
loguru==0.7.0
psutil==5.9.5
py-spy==0.3.14

# Utilidades
python-dotenv==1.0.0
tqdm==4.66.1
nltk==3.8.1 